Modif. dijkstra para pesos double

+++ gvgai/src/graph/Edge.java
@@ -4,9 +4,9 @@ public class Edge  {
     private final String id;
     private final Vertex source;
     private final Vertex destination;
-    private final int weight;
+    private final double weight;
 
-    public Edge(String id, Vertex source, Vertex destination, int weight) {
+    public Edge(String id, Vertex source, Vertex destination, double weight) {
             this.id = id;
             this.source = source;
             this.destination = destination;
@@ -23,7 +23,7 @@ public class Edge  {
     public Vertex getSource() {
             return source;
     }
-    public int getWeight() {
+    public double getWeight() {
             return weight;
     }
